pbkdf2: factor `Algorithm` and `Params` into modules
Moves these types out of the `phc` module, which should make it easier
to add MCF support (#747)
