+++ sha-crypt/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sha-crypt"
-version = "0.6.0-rc.1"
+version = "0.6.0-rc.2"
 description = """
 Pure Rust implementation of the SHA-crypt password hashing algorithm based on SHA-256/SHA-512
 as implemented by the POSIX crypt C library, including support for generating and verifying password
